cdn/check_cdn_line/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
dnspython==2.1.0
idna==3.3
prettytable==2.2.1
urllib3==1.26.7
wcwidth==0.2.5
requests==2.27.1
